+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Changed `choose` command for Scapegoat to take a space separated list rather than comma separated. ([#98](https://github.com/hjwylde/werewolf/issues/98))
 * Filtered `help` commands based on the current game. ([#94](https://github.com/hjwylde/werewolf/issues/94))
 * Added `--all` option to `help` commands. ([#94](https://github.com/hjwylde/werewolf/issues/94))
+* Added the Devoted Servant role. ([#47](https://github.com/hjwylde/werewolf/issues/47))
 
 #### v0.4.8.0